requirements.txt
pyserial
numpy
attrs
matplotlib
qtpy
qtawesome
pyqtgraph
qasync
scipy
pytest
nicelib
pyzmq
loguru
wrapt
nidaqmx
cffi
numba
skultrafast
PyQt5
h5py
pythonnet
lmfit
pytest-qt